requirements.txt
# Core requirements with relaxed versions
typing_extensions>=4.9.0
pytest>=8.0.0
pytest-mock>=3.12.0
PyYAML>=6.0.1
gym>=0.26.2
gymnasium>=0.28.1
ray[rllib]>=2.0.0
lz4>=4.3.2
pettingzoo>=1.22.0
sb3-contrib>=2.0.0
prettytable>=3.7.0
matplotlib>=3.7.0
networkx>=3.0.0
numpy>=1.24.0
torch>=2.0.0
paramiko>=3.0.0
sshtunnel>=0.4.0
tensorboard>=2.12.0
tensorboard-data-server>=0.7.0
tensorboardX>=2.6.0
pygame>=2.5.0
pygame-gui>=0.6.8
rich>=13.0.0
torch_geometric>=2.0.0

# New additions for LLM and Cybermonic agents
transformers>=4.35.0
trl>=0.7.1
openai
pandas>=2.0.0
scipy>=1.10.0
wandb>=0.15.0
tqdm>=4.64.0
scikit-image>=0.20.0
joblib>=1.1.0
dm_tree